Fix issues reported by PREfast static analysis tool (dotnet/coreclr#24577)
Commit migrated from https://github.com/dotnet/coreclr/commit/718598334310f6ad331ba2893ee4557beffaa74c

+++ src/clr/src/gc/env/gcenv.os.h
@@ -444,7 +444,7 @@ public:
     static void GetMemoryStatus(uint32_t* memory_load, uint64_t* available_physical, uint64_t* available_page_file);
 
     // Get size of an OS memory page
-    static uint32_t GetPageSize();
+    static size_t GetPageSize();
 
     //
     // Misc
